exocortex/requirements.txt
python-telegram-bot==21.4
Flask==3.0.2
python-dotenv==1.0.1
pinecone-client==4.1.0
google-generativeai==0.7.2
groq>=0.11.0
langchain==0.3.0
langchain-text-splitters==0.3.0
openai>=1.51.0
PyMuPDF==1.24.10
python-dateutil==2.9.0
requests==2.32.3
trafilatura==1.8.1
